+++ Package.swift
@@ -5,14 +5,14 @@ import PackageDescription
 let package = Package(
     name: "SweeterSwift",
     platforms: [
-        .iOS(.v9)
+        .iOS(.v9),
     ],
     products: [
-        .library( name: "SweeterSwift", targets: ["SweeterSwift"])
+        .library(name: "SweeterSwift", targets: ["SweeterSwift"]),
     ],
     dependencies: [],
     targets: [
-        .target(name: "SweeterSwift", dependencies: [], path: "Source")
+        .target(name: "SweeterSwift", dependencies: [], path: "Source"),
     ],
     swiftLanguageVersions: [.v5]
 )